// app_backend/src/Interfaces/Enums.ts
export enum validateIsUniqueInCollectionParams {
  INSERT,
  UPDATE
}

export const validLanguages = {
  pl: 'pl',
  en: 'en'
};

export enum ErrorTypes {
  NOTFOUND = 'notFound',
  INVALIDINPUT = 'invalidInput',
  UNAUTHORIZED = 'unauthorized',
  BADREQUEST = 'badRequest',
  REGULARERROR = 'regularError'
}